+++ current/samples/c/calculate_xy.c
@@ -20,24 +20,26 @@
 #include <math.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <inttypes.h>
 
-
-int main(int argc, char *argv[])
+int32_t main(int32_t argc, char *argv[])
 {
     urg_t urg;
-    long *data;
-    long max_distance;
-    long min_distance;
-    long time_stamp;
-    int i;
-    int n;
+    int32_t *data;
+    int32_t max_distance;
+    int32_t min_distance;
+    int32_t time_stamp;
+    int32_t i;
+    int32_t n;
 
-    if (open_urg_sensor(&urg, argc, argv) < 0) {
+    if (open_urg_sensor(&urg, argc, argv) < 0)
+    {
         return 1;
     }
 
-    data = (long *)malloc(urg_max_data_size(&urg) * sizeof(data[0]));
-    if (!data) {
+    data = (int32_t *)malloc(urg_max_data_size(&urg) * sizeof(data[0]));
+    if (!data)
+    {
         perror("urg_max_index()");
         return 1;
     }
@@ -46,7 +48,8 @@ int main(int argc, char *argv[])
     // \~english Gets measurement data
     urg_start_measurement(&urg, URG_DISTANCE, 1, 0, 1);
     n = urg_get_distance(&urg, data, &time_stamp);
-    if (n < 0) {
+    if (n < 0)
+    {
         printf("urg_get_distance: %s\n", urg_error(&urg));
         urg_close(&urg);
         return 1;
@@ -55,21 +58,23 @@ int main(int argc, char *argv[])
     // \~japanese X-Y 座標系の値を出力
     // \~english Outputs X-Y coordinates
     urg_distance_min_max(&urg, &min_distance, &max_distance);
-    for (i = 0; i < n; ++i) {
-        long distance = data[i];
+    for (i = 0; i < n; ++i)
+    {
+        int32_t distance = data[i];
         double radian;
-        long x;
-        long y;
+        int32_t x;
+        int32_t y;
 
-        if ((distance < min_distance) || (distance > max_distance)) {
+        if ((distance < min_distance) || (distance > max_distance))
+        {
             continue;
         }
 
         radian = urg_index2rad(&urg, i);
-        x = (long)(distance * cos(radian));
-        y = (long)(distance * sin(radian));
+        x = (int32_t)(distance * cos(radian));
+        y = (int32_t)(distance * sin(radian));
 
-        printf("%ld, %ld\n", x, y);
+        printf("%" PRId32 ", %" PRId32 "\n", x, y);
     }
     printf("\n");
