+++ current/samples/c/get_distance_intensity_io.c
@@ -11,69 +11,74 @@
 #include "urg_sensor.h"
 #include "urg_utils.h"
 #include "open_urg_sensor.h"
+#include <inttypes.h>
 #include <stdio.h>
 #include <stdlib.h>
 
-
-static void print_data(urg_t *urg, long data[], unsigned short intensity[],
-                       long io[], long time_stamp)
+static void print_data(urg_t *urg, int32_t data[], unsigned short intensity[],
+                       int32_t io[], int32_t time_stamp)
 {
-    enum {
+    enum
+    {
         IO_BIT_SIZE = 18,
     };
 
-    printf("timestamp: %ld\n", time_stamp);
+    printf("timestamp: %" PRId32 "\n", time_stamp);
 
     // \~japanese IO情報を表示
     // \~english Display IO information
-    unsigned long mask = (int)1 << (IO_BIT_SIZE - 1);
+    uint32_t mask = (uint32_t)1 << (IO_BIT_SIZE - 1);
     char in[IO_BIT_SIZE + 1];
     char out[IO_BIT_SIZE + 1];
-    int i;
-    for (i = 0; i < IO_BIT_SIZE; ++i) {
+    int32_t i;
+    for (i = 0; i < IO_BIT_SIZE; ++i)
+    {
         in[i] = (mask & io[0] ? '1' : '0');
         out[i] = (mask & io[1] ? '1' : '0');
         mask >>= 1;
     }
     in[IO_BIT_SIZE] = '\0';
     out[IO_BIT_SIZE] = '\0';
-    printf("input    : %s (%ld)\n", in, io[0]);
-    printf("output   : %s (%ld)\n", out, io[1]);
+    printf("input    : %s (%" PRId32 ")\n", in, io[0]);
+    printf("output   : %s (%" PRId32 ")\n", out, io[1]);
 
     // \~japanese 前方のデータのみを表示
     // \~english Shows only the front step
-    int front_index = urg_step2index(urg, 0);
-    printf("distance : %ld [mm]\n", data[front_index]);
-    printf("intensity: %ld [1]\n\n", intensity[front_index]);
+    int32_t front_index = urg_step2index(urg, 0);
+    printf("distance : %" PRId32 " [mm]\n", data[front_index]);
+    printf("intensity: %d [1]\n\n", intensity[front_index]);
 }
 
-
-int main(int argc, char *argv[])
+int32_t main(int32_t argc, char *argv[])
 {
-    enum {
+    enum
+    {
         CAPTURE_TIMES = 10,
     };
     urg_t urg;
-    int max_data_size;
-    long *io = malloc(2 * sizeof(long));
-    long *data = NULL;
+    int32_t max_data_size;
+    int32_t *io = malloc(2 * sizeof(int32_t));
+    int32_t *data = NULL;
     unsigned short *intensity = NULL;
-    long time_stamp;
-    int n;
-    int i;
+    int32_t time_stamp;
+    int32_t n;
+    int32_t i;
 
-    if (open_urg_sensor(&urg, argc, argv) < 0) {
+    if (open_urg_sensor(&urg, argc, argv) < 0)
+    {
         return 1;
     }
 
     max_data_size = urg_max_data_size(&urg);
-    data = (long *)malloc(max_data_size * sizeof(data[0]));
-    if (!data) {
+    data = (int32_t *)malloc(max_data_size * sizeof(data[0]));
+    if (!data)
+    {
         perror("urg_max_index()");
         return 1;
     }
     intensity = malloc(max_data_size * sizeof(intensity[0]));
-    if (!intensity) {
+    if (!intensity)
+    {
         perror("urg_max_index()");
         return 1;
     }
@@ -81,9 +86,11 @@ int main(int argc, char *argv[])
     // \~japanese データ取得
     // \~english Gets measurement data
     urg_start_measurement(&urg, URG_DISTANCE_INTENSITY_IO, URG_SCAN_INFINITY, 0);
-    for (i = 0; i < CAPTURE_TIMES; ++i) {
+    for (i = 0; i < CAPTURE_TIMES; ++i)
+    {
         n = urg_get_distance_intensity_io(&urg, data, intensity, io, &time_stamp);
-        if (n <= 0) {
+        if (n <= 0)
+        {
             printf("urg_get_distance_intensity_io: %s\n", urg_error(&urg));
             free(data);
             urg_close(&urg);
