+++ current/samples/c/get_multiecho_intensity.c
@@ -14,32 +14,33 @@
 #include "open_urg_sensor.h"
 #include <stdio.h>
 #include <stdlib.h>
+#include <inttypes.h>
 
-
-static void print_echo_data(long data[], unsigned short intensity[],
-                            int index)
+static void print_echo_data(int32_t data[], unsigned short intensity[],
+                            int32_t index)
 {
-    int i;
+    int32_t i;
 
     // [mm]
-    for (i = 0; i < URG_MAX_ECHO; ++i) {
-        printf("%ld, ", data[(URG_MAX_ECHO * index) + i]);
+    for (i = 0; i < URG_MAX_ECHO; ++i)
+    {
+        printf("%" PRId32 ", ", data[(URG_MAX_ECHO * index) + i]);
     }
 
     // [1]
-    for (i = 0; i < URG_MAX_ECHO; ++i) {
+    for (i = 0; i < URG_MAX_ECHO; ++i)
+    {
         printf("%d, ", intensity[(URG_MAX_ECHO * index) + i]);
     }
 }
 
-
 // \~japanese 距離、強度のデータを表示する
 // \~english Prints distance and intensity data
-static void print_data(urg_t *urg, long data[],
-                       unsigned short intensity[], int data_n, long time_stamp)
+static void print_data(urg_t *urg, int32_t data[],
+                       unsigned short intensity[], int32_t data_n, int32_t time_stamp)
 {
 #if 1
-    int front_index;
+    int32_t front_index;
 
     (void)data_n;
 
@@ -47,46 +48,49 @@ static void print_data(urg_t *urg, long data[],
     // \~english Shows only the front step
     front_index = urg_step2index(urg, 0);
     print_echo_data(data, intensity, front_index);
-    printf("%ld\n", time_stamp);
+    printf("%" PRId32 "\n", time_stamp);
 
 #else
     (void)urg;
-    int i;
+    int32_t i;
 
     // \~japanese 全てのデータを表示
     // \~english Prints the multiecho range/intensity values for all the measurement points
-    printf("# n = %d, time_stamp = %ld\n", data_n, time_stamp);
-    for (i = 0; i < data_n; ++i) {
+    printf("# n = %d, time_stamp = %" PRId32 "\n", data_n, time_stamp);
+    for (i = 0; i < data_n; ++i)
+    {
         print_echo_data(data, intensity, i);
         printf("\n");
     }
 #endif
 }
 
-
-int main(int argc, char *argv[])
+int32_t main(int32_t argc, char *argv[])
 {
-    enum {
+    enum
+    {
         CAPTURE_TIMES = 10,
     };
     urg_t urg;
-    int max_data_size;
-    long *data = NULL;
+    int32_t max_data_size;
+    int32_t *data = NULL;
     unsigned short *intensity = NULL;
-    long time_stamp;
-    int n;
-    int i;
+    int32_t time_stamp;
+    int32_t n;
+    int32_t i;
 
-    if (open_urg_sensor(&urg, argc, argv) < 0) {
+    if (open_urg_sensor(&urg, argc, argv) < 0)
+    {
         return 1;
     }
 
     max_data_size = urg_max_data_size(&urg);
-    data = (long *)malloc(max_data_size * 3 * sizeof(data[0]));
+    data = (int32_t *)malloc(max_data_size * 3 * sizeof(data[0]));
     intensity = (unsigned short *)malloc(max_data_size * 3 *
                                          sizeof(intensity[0]));
 
-    if (!data) {
+    if (!data)
+    {
         perror("urg_max_index()");
         return 1;
     }
@@ -94,9 +98,11 @@ int main(int argc, char *argv[])
     // \~japanese データ取得
     // \~english Gets measurement data
     urg_start_measurement(&urg, URG_MULTIECHO_INTENSITY, URG_SCAN_INFINITY, 0, 1);
-    for (i = 0; i < CAPTURE_TIMES; ++i) {
+    for (i = 0; i < CAPTURE_TIMES; ++i)
+    {
         n = urg_get_multiecho_intensity(&urg, data, intensity, &time_stamp);
-        if (n <= 0) {
+        if (n <= 0)
+        {
             printf("urg_get_multiecho_intensity: %s\n", urg_error(&urg));
             free(data);
             free(intensity);
